+++ src/Resource/AccountManageBilling.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Resource;
+
+class AccountManageBilling extends BaseResource
+{
+    public string $url;
+}